The scheduled task to update the free-credits spreadsheet was not performant. For reference: https://docs.google.com/document/d/12pwsl_xT1THr-e2K-Dcx35s3baWk_11QQ02VG3eEg2E/edit?usp=sharing (access allowed to Broad internal only)

This PR changes it such that:
* instead of one big query to Elasticsearch (hardcoded limit: 5000 documents), we make multiple paginated [scroll queries](https://www.elastic.co/guide/en/elasticsearch/reference/current/search-request-scroll.html) to get all documents.
* instead of N small queries to Thurloe, where N=number of free trial users, we batch the queries to Thurloe into groups of 100 users.
* more resilience to/error-handling of missing data when generating the spreadsheet cells

Some rough analysis of how this would change query load in production:

service | Before  PR | After PR
--- | ---------- | ---------- |
Elasticsearch |  1 query, returning 1649 documents | 7 queries, returning up to 250 documents
Thurloe | 1579 queries, 1 user per query | 40 queries, up to 40 users per query
 Thurloe | all keys per user, roughly 20 | 13 keys per  user

-----
Changes since this PR was approved (8dd899080eeabb704275a1cfc23317579e175e30):
1. Changed the order of `key` and `userid` query params in the request to Thurloe. This is purely cosmetic: since we expect the `key`s to be the same for all requests but the `userid`s to  change, I  wanted the static params to come first.
2. Changed the batch size from 100 to 40. It was necessary to reduce this below  ~50 to address url-length restrictions (I originally miscalculated when choosing 100). From empirical perf testing, 40 seemed the sweet spot.
3. Ran a bunch of perf tests at multiple levels of concurrency and multiple query batch sizes. Landed on 40 as the best batch size. I have this data if anyone wants to review.
4. Stood up this version of orch in perf and watched the spreadsheet-update process repeatedly at a batch size of 40. It performed well and the requests to Thurloe were all fast - most under 100ms with a few outliers.

-----

Have you read [CONTRIBUTING.md](../CONTRIBUTING.md) lately? If not, do that first.

I, the developer opening this PR, do solemnly pinky swear that:

- [ ] I've followed [the instructions](https://github.com/broadinstitute/firecloud-orchestration/blob/develop/CONTRIBUTING.md#api-changes) if I've made any changes to the API, _especially_ if they're breaking changes
- [ ] I've updated the RC_XXX release ticket with any manual steps required to release this change
- [ ] I've updated the [FISMA documentation](https://github.com/broadinstitute/firecloud-orchestration/blob/develop/CONTRIBUTING.md#fisma-documentation-changes) if I've made any security-related changes, including auth, encryption, or auditing

In all cases:

- [ ] Get two thumbsworth of review and PO signoff if necessary
- [ ] Verify all tests go green
- [ ] Squash and merge; you can delete your branch after this **unless it's for a hotfix**. In that case, don't delete it!
- [ ] Test this change deployed correctly and works on dev environment after deployment