@davidangb 2019-05-21
thanks @rtitle ! Next step after this is perf testing; currently bottlenecked on access to the perf env. I'm hoping that the only tweak to this code as a result of perf testing is the batch size -  currently 100 - for queries to Thurloe, so I felt safe  putting this PR up for code  review.