@rtitle 2019-05-31
The latest changes look good to me, one question about whether the magic numbers (250, 40, etc) should live in a config file. My approval still stands though. 👍