@davidangb 2019-06-03
@rtitle thanks for the review! I'm not putting these in config because I'm not sure if we'll ever change them or use environment-specific values, so I don't want to add overhead/confusion. But I did add some explicit code comments in those areas.